Maria: edit AnularLineaPedido

+++ public_html/app/Controllers/Pedidos.php
@@ -136,11 +136,13 @@ class Pedidos extends BaseController
 		$pedidoModel = new Pedidos_model($db);
 
 		// Validación básica de datos
-		if (!$this->validate([
-			'id_cliente' => 'required',
-			'fecha_entrada' => 'required',
-			'fecha_entrega' => 'required',
-		])) {
+		if (
+			!$this->validate([
+				'id_cliente' => 'required',
+				'fecha_entrada' => 'required',
+				'fecha_entrega' => 'required',
+			])
+		) {
 			return redirect()->back()->with('error', 'Faltan datos obligatorios');
 		}
 
@@ -180,8 +182,8 @@ class Pedidos extends BaseController
 	public function edit($id_pedido)
 	{
 		$this->addBreadcrumb('Inicio', base_url('/'));
-        $this->addBreadcrumb('Pedidos', base_url('/pedidos/enmarcha'));
-        $this->addBreadcrumb('Editar Pedido');
+		$this->addBreadcrumb('Pedidos', base_url('/pedidos/enmarcha'));
+		$this->addBreadcrumb('Editar Pedido');
 		helper('controlacceso');
 		$session = session();
 		$data = datos_user();
@@ -221,22 +223,24 @@ class Pedidos extends BaseController
 		$data = usuario_sesion();
 		$db = db_connect($data['new_db']);
 		$pedidoModel = new Pedidos_model($db);
-	
+
 		// Validación básica de datos
-		if (!$this->validate([
-			'id_cliente' => 'required',
-			'fecha_entrada' => 'required',
-			'fecha_entrega' => 'required',
-		])) {
+		if (
+			!$this->validate([
+				'id_cliente' => 'required',
+				'fecha_entrada' => 'required',
+				'fecha_entrega' => 'required',
+			])
+		) {
 			return redirect()->back()->with('error', 'Faltan datos obligatorios');
 		}
-	
+
 		// Obtener el pedido actual para mantener su estado
 		$pedido = $pedidoModel->find($id_pedido);
 		if (!$pedido) {
 			return redirect()->back()->with('error', 'Pedido no encontrado');
 		}
-	
+
 		// Preparar los datos para actualizar el pedido
 		$updateData = [
 			'id_cliente' => $this->request->getPost('id_cliente'),
@@ -246,10 +250,10 @@ class Pedidos extends BaseController
 			'fecha_entrega' => $this->request->getPost('fecha_entrega'),
 			'observaciones' => $this->request->getPost('observaciones'),
 		];
-	
+
 		// Mantener el estado original del pedido, no modificarlo
 		$updateData['estado'] = $pedido->estado; // Usar notación de objeto ->
-	
+
 		// Actualizar el pedido
 		if ($pedidoModel->update($id_pedido, $updateData)) {
 			return redirect()->to(base_url('pedidos/edit/' . $id_pedido))->with('success', 'Pedido actualizado correctamente');
@@ -257,7 +261,7 @@ class Pedidos extends BaseController
 			return redirect()->back()->with('error', 'No se pudo actualizar el pedido');
 		}
 	}
-	
+
 
 	function imprimir_parte($row)
 	{
@@ -304,6 +308,7 @@ class Pedidos extends BaseController
 		$builder = $db->table('linea_pedidos');
 		$builder->selectSum('total_linea', 'suma_total');
 		$builder->where('id_pedido', $id_pedido);
+		$builder->where('estado !=', 6);
 		$query = $builder->get();
 		$resultado = $query->getRow();
 		$totalPedido = $resultado->suma_total ?? 0;
@@ -311,6 +316,7 @@ class Pedidos extends BaseController
 		$pedidoModel->update($id_pedido, ['total_pedido' => $totalPedido]);
 		return $totalPedido;
 	}
+
 	public function entregar($id_pedido)
 	{
 		$Lineaspedido_model = model('App\Models\Lineaspedido_model');
@@ -340,9 +346,11 @@ class Pedidos extends BaseController
 		$data = usuario_sesion();
 		$db = db_connect($data['new_db']);
 		$lineaspedidoModel = new LineaPedido($db);
-		if (!$this->validate([
-			'id_producto' => 'required',
-		])) {
+		if (
+			!$this->validate([
+				'id_producto' => 'required',
+			])
+		) {
 			return redirect()->back()->with('error', 'El producto es obligatorio.');
 		}
 		$fecha_entrada = $this->request->getPost('fecha_entrada') ?: date('Y-m-d');
@@ -426,20 +434,20 @@ class Pedidos extends BaseController
 		$estadoModel = new EstadoModel($db);
 		$lineaPedidoModel = new LineaPedido($db);
 		$linea_pedido = $lineaPedidoModel->find($id_lineapedido);
-	
+
 		if (!$linea_pedido) {
 			return $this->response->setStatusCode(404, 'Línea de pedido no encontrada');
 		}
-	
+
 		// Verificar si el estado de la línea es "en cola"
 		$isEstadoEnCola = ($linea_pedido['estado'] === 'en cola');
-	
+
 		// Pasar datos a la vista
 		$data['productos'] = $productosModel->findAll();
 		$data['estados'] = $estadoModel->findAll();
 		$data['linea_pedido'] = $linea_pedido;
 		$data['isEstadoEnCola'] = $isEstadoEnCola;  // Variable adicional para controlar la visibilidad
-	
+
 		// Renderizar la vista dependiendo de si es AJAX o no
 		if ($this->request->isAJAX()) {
 			return view('editLineaPedido', $data);
@@ -447,7 +455,7 @@ class Pedidos extends BaseController
 			return redirect()->back()->with('error', 'Acción no permitida');
 		}
 	}
-	
+
 	public function mostrarFormularioAddLineaPedido($id_pedido)
 	{
 		$data = usuario_sesion();
@@ -465,23 +473,61 @@ class Pedidos extends BaseController
 	{
 		$data = usuario_sesion();
 		$db = db_connect($data['new_db']);
+		$session = session();
+		$session_data = $session->get('logged_in');
+		$nivel_acceso = $session_data['nivel'];
+
 		$lineaPedidoModel = new LineaPedido($db);
 		$procesosPedidoModel = new ProcesosPedido($db);
 		$linea = $lineaPedidoModel->where('id_lineapedido', $id_lineapedido)->first();
+
 		if (!$linea) {
 			return redirect()->back()->with('error', 'Línea no encontrada');
 		}
+
 		$id_pedido = $linea['id_pedido'];
+
+		if ($nivel_acceso != 9) {
+			return $this->anularLinea($id_lineapedido, $id_pedido);
+		}
+
 		$db->transStart();
 		$procesosPedidoModel->where('id_linea_pedido', $id_lineapedido)->delete();
 		$lineaPedidoModel->delete($id_lineapedido);
 		$db->transComplete();
+
 		if ($db->transStatus() === false) {
 			return redirect()->back()->with('error', 'No se pudo eliminar la línea del pedido');
 		}
-		$this->logAction('Pedidos', 'Elimina Linea pedido, ID: ' . $id_lineapedido, []);
+
+		$this->logAction('Pedidos', 'Elimina Línea pedido, ID: ' . $id_lineapedido, []);
+
 		return redirect()->to(base_url('pedidos/edit/' . $id_pedido))->with('success', 'Línea del pedido y procesos asociados eliminados correctamente');
 	}
+
+	public function anularLinea($id_lineapedido, $id_pedido)
+	{
+		$data = usuario_sesion();
+		$db = db_connect($data['new_db']);
+		$lineaPedidoModel = new LineaPedido($db);
+
+		$linea = $lineaPedidoModel->where('id_lineapedido', $id_lineapedido)->first();
+
+		if (!$linea) {
+			return redirect()->to(base_url('pedidos/edit/' . $id_pedido))->with('error', 'La línea de pedido no existe.');
+		}
+
+		$update = $lineaPedidoModel->update($id_lineapedido, ['estado' => 6]);
+
+		if ($update) {
+			$totalPedido = $this->actualizarTotalPedido($id_pedido);
+			$this->logAction('Pedidos', 'Anula Línea pedido, ID: ' . $id_lineapedido, []);
+			return redirect()->to(base_url('pedidos/edit/' . $id_pedido))->with('success', 'Línea de pedido anulada correctamente y total del pedido actualizado. Total: ' . $totalPedido);
+		} else {
+			return redirect()->to(base_url('pedidos/edit/' . $id_pedido))->with('error', 'No se pudo anular la línea de pedido');
+		}
+	}
+
 
 	public function actualizarEstadoPedido($id_pedido)
 	{
